+++ packages/local-prover/src/joinsplit.ts
@@ -5,7 +5,9 @@ import {
   JoinSplitInputs,
   JoinSplitProofWithPublicSignals,
   JoinSplitProver,
+  StealthAddressTrait,
   VerifyingKey,
+  decomposeCompressedPoint,
 } from "@nocturne-xyz/sdk";
 
 export class WasmJoinSplitProver implements JoinSplitProver {
@@ -35,9 +37,14 @@ export class WasmJoinSplitProver implements JoinSplitProver {
       merkleProofB,
       newNoteA,
       newNoteB,
-      encRandomness,
+      refundAddr,
       encodedAssetAddrWithSignBits,
+      encodedAssetId,
     } = inputs;
+
+    const [, refundAddrH1CompressedY] = decomposeCompressedPoint(refundAddr.h1);
+    const [, refundAddrH2CompressedY] = decomposeCompressedPoint(refundAddr.h2);
+    const decompressedRefundAddr = StealthAddressTrait.decompress(refundAddr);
 
     const signals = {
       userViewKey: vk,
@@ -45,11 +52,21 @@ export class WasmJoinSplitProver implements JoinSplitProver {
       spendPubKey: spendPk,
       userViewKeyNonce: vkNonce,
 
-      encodedAssetId: oldNoteA.encodedAssetId,
       operationDigest,
 
       c,
       z,
+
+      encodedAssetAddrWithSignBits,
+      encodedAssetId,
+
+      refundAddrH1CompressedY,
+      refundAddrH2CompressedY,
+
+      refundAddrH1X: decompressedRefundAddr.h1X,
+      refundAddrH1Y: decompressedRefundAddr.h1Y,
+      refundAddrH2X: decompressedRefundAddr.h2X,
+      refundAddrH2Y: decompressedRefundAddr.h2Y,
 
       oldNoteAOwnerH1X: oldNoteA.owner.h1X,
       oldNoteAOwnerH1Y: oldNoteA.owner.h1Y,
@@ -75,9 +92,6 @@ export class WasmJoinSplitProver implements JoinSplitProver {
 
       receiverCanonAddr: [newNoteB.owner.h2X, newNoteB.owner.h2Y],
       newNoteBValue: newNoteB.value,
-
-      encRandomness,
-      encodedAssetAddrWithSignBits,
     };
 
     const proof = await snarkjs.groth16.fullProve(
@@ -86,6 +100,10 @@ export class WasmJoinSplitProver implements JoinSplitProver {
       this.zkeyPath
     );
 
+    // ensure publicSignals are BigInts
+    proof.publicSignals = proof.publicSignals.map((val: any) =>
+      BigInt(val as string)
+    );
     return proof;
   }