+++ packages/sdk/src/prepareOperation.ts
@@ -11,13 +11,15 @@ import {
   Asset,
   AssetTrait,
   PreSignOperation,
+  SENDER_COMMITMENT_DOMAIN_SEPARATOR,
 } from "./primitives";
 import {
   NocturneViewer,
   CanonAddress,
   StealthAddressTrait,
   encryptNote,
-  randomBigInt,
+  randomFr,
+  CompressedStealthAddress,
 } from "./crypto";
 import { MerkleProofInput } from "./proof";
 import {
@@ -28,6 +30,7 @@ import {
   groupByArr,
 } from "./utils";
 import { SparseMerkleProver } from "./SparseMerkleProver";
+import { poseidonBN } from "@nocturne-xyz/crypto-utils";
 
 export const __private = {
   gatherNotes,
@@ -47,6 +50,11 @@ export async function prepareOperation(
   const encodedRefundAssets = refundAssets.map(AssetTrait.encode);
   const encodedGasAsset = AssetTrait.encode(opRequest.gasAsset);
 
+  // if refundAddr is not set, generate a random one
+  const refundAddr = StealthAddressTrait.compress(
+    opRequest.refundAddr ?? deps.viewer.generateRandomStealthAddress()
+  );
+
   // prepare joinSplits
   let joinSplits: PreSignJoinSplit[] = [];
   const usedMerkleIndices = new Set<number>();
@@ -55,6 +63,7 @@ export async function prepareOperation(
     const newJoinSplits = await prepareJoinSplits(
       deps,
       joinSplitRequest,
+      refundAddr,
       usedMerkleIndices
     );
 
@@ -74,11 +83,6 @@ export async function prepareOperation(
     AssetTrait.encodedAssetToString(joinSplit.encodedAsset)
   ).flat();
 
-  // if refundAddr is not set, generate a random one
-  const refundAddr = StealthAddressTrait.compress(
-    opRequest.refundAddr ?? deps.viewer.generateRandomStealthAddress()
-  );
-
   // construct op.
   const op: PreSignOperation = {
     ...opRequest,
@@ -97,6 +101,7 @@ export async function prepareOperation(
 async function prepareJoinSplits(
   { db, viewer, merkle }: PrepareOperationDeps,
   joinSplitRequest: JoinSplitRequest,
+  refundAddr: CompressedStealthAddress,
   alreadyUsedNoteMerkleIndices: Set<number> = new Set()
 ): Promise<PreSignJoinSplit[]> {
   const notes = await gatherNotes(
@@ -121,6 +126,7 @@ async function prepareJoinSplits(
     notes,
     paymentAmount,
     amountToReturn,
+    refundAddr,
     receiver
   );
 }
@@ -200,12 +206,13 @@ async function getJoinSplitsFromNotes(
   notes: IncludedNote[],
   paymentAmount: bigint,
   amountLeftOver: bigint,
+  refundAddr: CompressedStealthAddress,
   receiver?: CanonAddress
 ): Promise<PreSignJoinSplit[]> {
   // add a dummy note if there are an odd number of notes.
   if (notes.length % 2 == 1) {
     const newAddr = viewer.generateRandomStealthAddress();
-    const nonce = randomBigInt();
+    const nonce = randomFr();
     notes.push({
       owner: newAddr,
       nonce,
@@ -235,6 +242,7 @@ async function getJoinSplitsFromNotes(
       noteB,
       paymentAmount,
       amountToReturn,
+      refundAddr,
       receiver
     );
 
@@ -251,6 +259,7 @@ async function makeJoinSplit(
   oldNoteB: IncludedNote,
   paymentAmount: bigint,
   amountToReturn: bigint,
+  refundAddr: CompressedStealthAddress,
   receiver?: CanonAddress
 ): Promise<PreSignJoinSplit> {
   const sender = viewer.canonicalAddress();
@@ -285,8 +294,8 @@ async function makeJoinSplit(
   const newNoteACommitment = NoteTrait.toCommitment(newNoteA);
   const newNoteBCommitment = NoteTrait.toCommitment(newNoteB);
 
-  const newNoteAEncrypted = encryptNote(sender, newNoteA);
-  const newNoteBEncrypted = encryptNote(receiver, newNoteB);
+  const newNoteAEncrypted = encryptNote(sender, { ...newNoteA, sender });
+  const newNoteBEncrypted = encryptNote(receiver, { ...newNoteB, sender });
 
   const membershipProof = merkle.getProof(oldNoteA.merkleIndex);
   const commitmentTreeRoot = membershipProof.root;
@@ -318,6 +327,15 @@ async function makeJoinSplit(
     };
   }
 
+  // commit to the sender's canonical address
+  const senderCanonAddr = viewer.canonicalAddress();
+  const senderCommitment = poseidonBN([
+    SENDER_COMMITMENT_DOMAIN_SEPARATOR,
+    senderCanonAddr.x,
+    senderCanonAddr.y,
+    newNoteB.nonce,
+  ]);
+
   return {
     receiver,
     encodedAsset,
@@ -338,5 +356,8 @@ async function makeJoinSplit(
     newNoteBCommitment,
     merkleProofA,
     merkleProofB,
+
+    senderCommitment,
+    refundAddr,
   };
 }