+++ packages/sdk/src/primitives/note.ts
@@ -1,9 +1,10 @@
-import { StealthAddressTrait, StealthAddress } from "../crypto";
+import { StealthAddressTrait, StealthAddress, CanonAddress } from "../crypto";
 import { Asset, AssetTrait, EncodedAsset } from "./asset";
-import { poseidonBN } from "@nocturne-xyz/circuit-utils";
+import { poseidonBN } from "@nocturne-xyz/crypto-utils";
 import { sha256 } from "js-sha256";
 import { bigintToBEPadded } from "../utils";
 import { EncryptedNote } from "./types";
+import { bigintFromBEBytes } from "../utils/bits";
 
 export interface Note {
   owner: StealthAddress;
@@ -14,6 +15,10 @@ export interface Note {
 
 export interface IncludedNote extends Note {
   merkleIndex: number;
+}
+
+export interface NoteWithSender extends Note {
+  sender: CanonAddress;
 }
 
 export interface IncludedNoteWithNullifier extends IncludedNote {
@@ -32,6 +37,24 @@ export interface EncodedNote {
   encodedAssetId: bigint;
   value: bigint;
 }
+
+// compact note byte-serialization format (181 bytes)
+// all bigints are represented in big-endian encoding
+// 1. owner: 64 bytes (h1Compressed: 32 bytes, h2Compressed: 32 bytes)
+// 2. asset: 53 bytes (AssetTrait.serializeCompact(asset))
+// 3. nonce: 32 bytes
+// 4. value: 32 bytes
+const NOTE_COMPACT_SERIALIZE_BYTES = 181;
+const NOTE_COMPACT_SERIALIZE_OFFSETS = {
+  owner: {
+    h1: 0,
+    h2: 32,
+  },
+  asset: 64,
+  nonce: 117,
+  value: 149,
+  end: 181,
+};
 
 export class NoteTrait {
   static toCommitment<N extends Note>(note: N): bigint {
@@ -85,7 +108,7 @@ export class NoteTrait {
   static isEncryptedNote<N extends Note, E extends EncryptedNote>(
     note: N | E
   ): boolean {
-    return Object.hasOwn(note, "encappedKey");
+    return Object.hasOwn(note, "encapsulatedSecretBytes");
   }
 
   static sha256<N extends Note>(note: N): number[] {
@@ -124,6 +147,63 @@ export class NoteTrait {
   }
 
   static toNote<N extends Note>({ owner, nonce, asset, value }: N): Note {
+    return { owner, nonce, asset, value };
+  }
+
+  // see above for serialization format
+  static serializeCompact<N extends Note>(note: N): Uint8Array {
+    const buf = new Uint8Array(NOTE_COMPACT_SERIALIZE_BYTES);
+
+    const { owner, nonce, asset, value } = note;
+    const { h1, h2 } = StealthAddressTrait.compress(owner);
+
+    buf.set(bigintToBEPadded(h1, 32), NOTE_COMPACT_SERIALIZE_OFFSETS.owner.h1);
+    buf.set(bigintToBEPadded(h2, 32), NOTE_COMPACT_SERIALIZE_OFFSETS.owner.h2);
+    buf.set(
+      AssetTrait.serializeCompact(asset),
+      NOTE_COMPACT_SERIALIZE_OFFSETS.asset
+    );
+    buf.set(bigintToBEPadded(nonce, 32), NOTE_COMPACT_SERIALIZE_OFFSETS.nonce);
+    buf.set(bigintToBEPadded(value, 32), NOTE_COMPACT_SERIALIZE_OFFSETS.value);
+
+    return buf;
+  }
+
+  static deserializeCompact(buf: Uint8Array): Note {
+    const h1 = bigintFromBEBytes(
+      buf.slice(
+        NOTE_COMPACT_SERIALIZE_OFFSETS.owner.h1,
+        NOTE_COMPACT_SERIALIZE_OFFSETS.owner.h2
+      )
+    );
+    const h2 = bigintFromBEBytes(
+      buf.slice(
+        NOTE_COMPACT_SERIALIZE_OFFSETS.owner.h2,
+        NOTE_COMPACT_SERIALIZE_OFFSETS.asset
+      )
+    );
+    const compressedOwner = { h1, h2 };
+    const owner = StealthAddressTrait.decompress(compressedOwner);
+
+    const asset = AssetTrait.deserializeCompact(
+      buf.slice(
+        NOTE_COMPACT_SERIALIZE_OFFSETS.asset,
+        NOTE_COMPACT_SERIALIZE_OFFSETS.nonce
+      )
+    );
+    const nonce = bigintFromBEBytes(
+      buf.slice(
+        NOTE_COMPACT_SERIALIZE_OFFSETS.nonce,
+        NOTE_COMPACT_SERIALIZE_OFFSETS.value
+      )
+    );
+    const value = bigintFromBEBytes(
+      buf.slice(
+        NOTE_COMPACT_SERIALIZE_OFFSETS.value,
+        NOTE_COMPACT_SERIALIZE_OFFSETS.end
+      )
+    );
+
     return { owner, nonce, asset, value };
   }