+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased
 
+- change `EncryptedNote` to be an alias of `HybridCiphertext` and remove all other fields (including owner)
+  - includes (many) cascading changes through sync, op preparation, and witness generation code
+- add `refundAddr` circuit input
+- replace `encSenderCanonAddr` with `senderCommitment`
 - add contract.ts utils method for concatenating contract address and function selector
 - fix eip712 deposit request typehash ordering
 - remove all usage of winston logger except for in sync adapters (so we can debug test actor and snap)
@@ -167,7 +171,7 @@
   - make exports explicit
   - clean up imports so that they don't depend on internal structure of other modules
 - merge `NocturnePrivKey` and `NocturneSigner`
-- replace `circomlibjs` and `ffjavascript` with `@nocturne-xyz/circuit-utils`
+- replace `circomlibjs` and `ffjavascript` with `@nocturne-xyz/crypto-utils`
 - replace `NocturneContext.test.ts` with unit tests for each of the newly-split-out parts
 - disembowel `NocturneContext`, reducing it to a convencience wrapper for all private state
 - add new module / function `proveOperation` that separates the proving logic form `NocturneContext`